Migrate `MavenDependencyResolveIntegrationTest` to new test fixtures
Issue #3394
