+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/maven/MavenDependencyResolveIntegrationTest.groovy
@@ -15,40 +15,49 @@
  */
 package org.gradle.integtests.resolve.maven
 
-import org.gradle.integtests.fixtures.AbstractDependencyResolutionTest
-import org.gradle.integtests.fixtures.resolve.ResolveTestFixture
-import org.gradle.util.Requires
-import org.gradle.util.TestPrecondition
+import org.gradle.integtests.fixtures.GradleMetadataResolveRunner
+import org.gradle.integtests.fixtures.RequiredFeature
+import org.gradle.integtests.fixtures.RequiredFeatures
+import org.gradle.integtests.resolve.AbstractModuleDependencyResolveTest
 
-class MavenDependencyResolveIntegrationTest extends AbstractDependencyResolutionTest {
-    def resolve = new ResolveTestFixture(buildFile)
+@RequiredFeatures(
+    @RequiredFeature(feature = GradleMetadataResolveRunner.REPOSITORY_TYPE, value = "maven")
+)
+class MavenDependencyResolveIntegrationTest extends AbstractModuleDependencyResolveTest {
 
-    def setup() {
-        resolve.prepare()
-        settingsFile << """
-            rootProject.name = 'testproject'
-        """
-    }
+    String getRootProjectName() { 'testproject' }
 
     def "dependency includes main artifact and runtime dependencies of referenced module"() {
         given:
-        def other = mavenRepo.module("org.gradle", "other", "preview-1").publish()
-        mavenRepo.module("org.gradle", "test", "1.45")
-            .dependsOn(other)
-            .artifact(classifier: 'classifier') // ignored
-            .publish()
+        repository {
+            'org.gradle:other:preview-1'()
+            'org.gradle:test:1.45' {
+                dependsOn 'org.gradle:other:preview-1'
+                withModule {
+                    artifact(classifier: 'classifier') // ignored
+                }
+            }
+        }
 
         and:
-
         buildFile << """
 group = 'org.gradle'
 version = '1.0'
-repositories { maven { url "${mavenRepo.uri}" } }
-configurations { compile }
 dependencies {
-    compile "org.gradle:test:1.45"
+    conf "org.gradle:test:1.45"
 }
 """
+
+        repositoryInteractions {
+            'org.gradle:test:1.45' {
+                expectGetMetadata()
+                expectGetArtifact()
+            }
+            'org.gradle:other:preview-1' {
+                expectGetMetadata()
+                expectGetArtifact()
+            }
+        }
 
         expect:
         succeeds "checkDep"
@@ -63,23 +72,42 @@ dependencies {
 
     def "dependency that references a classifier includes the matching artifact only plus the runtime dependencies of referenced module"() {
         given:
-        def other = mavenRepo.module("org.gradle", "other", "preview-1").publish()
-        mavenRepo.module("org.gradle", "test", "1.45")
-            .dependsOn(other)
-            .artifact(classifier: 'classifier')
-            .artifact(classifier: 'some-other') // ignored
-            .publish()
+        repository {
+            'org.gradle' {
+                'other' {
+                    'preview-1'()
+                }
+                'test' {
+                    '1.45' {
+                        dependsOn 'org.gradle:other:preview-1'
+                        withModule {
+                            artifact(classifier: 'classifier')
+                            artifact(classifier: 'some-other') // ignored
+                        }
+                    }
+                }
+            }
+        }
 
         and:
         buildFile << """
 group = 'org.gradle'
 version = '1.0'
-repositories { maven { url "${mavenRepo.uri}" } }
-configurations { compile }
 dependencies {
-    compile "org.gradle:test:1.45:classifier"
+    conf "org.gradle:test:1.45:classifier"
 }
 """
+
+        repositoryInteractions {
+            'org.gradle:test:1.45' {
+                expectGetMetadata()
+                expectGetArtifact(classifier: 'classifier')
+            }
+            'org.gradle:other:preview-1' {
+                expectGetMetadata()
+                expectGetArtifact()
+            }
+        }
 
         expect:
         succeeds "checkDep"
@@ -95,20 +123,28 @@ dependencies {
 
     def "dependency that references an artifact includes the matching artifact only plus the runtime dependencies of referenced module"() {
         given:
-        def other = mavenRepo.module("org.gradle", "other", "preview-1").publish()
-        mavenRepo.module("org.gradle", "test", "1.45")
-            .dependsOn(other)
-            .artifact(type: 'aar', classifier: 'classifier')
-            .publish()
+        repository {
+            'org.gradle' {
+                'other' {
+                    'preview-1'()
+                }
+                'test' {
+                    '1.45' {
+                        dependsOn 'org.gradle:other:preview-1'
+                        withModule {
+                            artifact(type: 'aar', classifier: 'classifier')
+                        }
+                    }
+                }
+            }
+        }
 
         and:
         buildFile << """
 group = 'org.gradle'
 version = '1.0'
-repositories { maven { url "${mavenRepo.uri}" } }
-configurations { compile }
 dependencies {
-    compile ("org.gradle:test:1.45") {
+    conf ("org.gradle:test:1.45") {
         artifact {
             name = 'test'
             type = 'aar'
@@ -117,6 +153,16 @@ dependencies {
     }
 }
 """
+        repositoryInteractions {
+            'org.gradle:test:1.45' {
+                expectGetMetadata()
+                expectGetArtifact(type: 'aar', classifier: 'classifier')
+            }
+            'org.gradle:other:preview-1' {
+                expectGetMetadata()
+                expectGetArtifact()
+            }
+        }
 
         expect:
         succeeds "checkDep"
@@ -130,23 +176,32 @@ dependencies {
         }
     }
 
+    @RequiredFeatures(
+        // only available with Maven metadata: Gradle metadata does not support "optional"
+        @RequiredFeature(feature = GradleMetadataResolveRunner.GRADLE_METADATA, value = "false")
+    )
     def "does not include optional dependencies of maven module"() {
         given:
-        def notRequired = mavenRepo.module("org.gradle", "i-do-not-exist", "1.0")
-        mavenRepo.module("org.gradle", "test", "1.45")
-            .dependsOn(notRequired, optional: true)
-            .dependsOn(notRequired, optional: true, scope: 'runtime')
-            .publish()
-
+        repository {
+            'org.gradle:test:1.45' {
+                dependsOn group:'org.gradle', artifact:'i-do-not-exist', version:'1.45', optional: 'true'
+                dependsOn group:'org.gradle', artifact:'i-do-not-exist', version:'1.45', optional: 'true', scope: 'runtime'
+            }
+        }
         and:
 
         buildFile << """
-repositories { maven { url "${mavenRepo.uri}" } }
-configurations { compile }
 dependencies {
-    compile "org.gradle:test:1.45"
+    conf "org.gradle:test:1.45"
 }
 """
+
+        repositoryInteractions {
+            'org.gradle:test:1.45' {
+                expectGetMetadata()
+                expectGetArtifact()
+            }
+        }
 
         expect:
         succeeds "checkDep"
@@ -157,91 +212,39 @@ dependencies {
         }
     }
 
-    @Requires(TestPrecondition.ONLINE)
-    def "resolves dependencies on real projects"() {
-        // Real but ancient projects
-        // Hibernate core brings in conflicts, exclusions and root poms
-        // Add a direct dependency on an earlier version of commons-collection than required by hibernate core
-        // Logback classic depends on a later version of slf4j-api than required by hibernate core
-
-        given:
-        buildFile << """
-${mavenCentralRepository()}
-
-configurations {
-    compile
-}
-
-dependencies {
-    compile "commons-collections:commons-collections:3.0"
-    compile "ch.qos.logback:logback-classic:0.9.30"
-    compile "org.hibernate:hibernate-core:3.6.7.Final"
-}
-
-task check {
-    doLast {
-        def compile = configurations.compile
-
-        def filteredDependencies = compile.resolvedConfiguration.getFirstLevelModuleDependencies({ it.name == 'logback-classic' } as Spec)
-        assert filteredDependencies.collect { it.name } == [
-            'ch.qos.logback:logback-classic:0.9.30'
-        ]
-
-        def filteredFiles = compile.resolvedConfiguration.getFiles({ it.name == 'logback-classic' } as Spec)
-        assert filteredFiles.collect { it.name } == [
-            'logback-classic-0.9.30.jar',
-            'slf4j-api-1.6.2.jar',
-            'logback-core-0.9.30.jar'
-        ]
-    }
-}
-"""
-
-        expect:
-        succeeds "check", "checkDep"
-        resolve.expectGraph {
-            root(':', ':testproject:') {
-                module('ch.qos.logback:logback-classic:0.9.30') {
-                    module('ch.qos.logback:logback-core:0.9.30')
-                    module('org.slf4j:slf4j-api:1.6.2')
-                }
-                module('org.hibernate:hibernate-core:3.6.7.Final') {
-                    module('org.hibernate:hibernate-commons-annotations:3.2.0.Final') {
-                        edge('org.slf4j:slf4j-api:1.5.8', 'org.slf4j:slf4j-api:1.6.2') {
-                            byConflictResolution()
-                        }
-                    }
-                    module('org.hibernate.javax.persistence:hibernate-jpa-2.0-api:1.0.1.Final')
-                    module('antlr:antlr:2.7.6')
-                    module('dom4j:dom4j:1.6.1')
-                    module('javax.transaction:jta:1.1')
-                    module('commons-collections:commons-collections:3.1')
-                    edge('org.slf4j:slf4j-api:1.6.1', 'org.slf4j:slf4j-api:1.6.2') {
-                        byConflictResolution()
-                    }
-                }
-                edge('commons-collections:commons-collections:3.0', 'commons-collections:commons-collections:3.1') {
-                    byConflictResolution()
-                }
-            }
-        }
-    }
-
+    @RequiredFeatures(
+        // exclusions are not supported by Gradle metadata
+        @RequiredFeature(feature = GradleMetadataResolveRunner.GRADLE_METADATA, value = "false")
+    )
     def "dependency with wildcard exclusions is treated as non-transitive"() {
         given:
-        def thirdLevel = mavenRepo.module("org.gradle", "thirdLevel", "1.0").publish()
-        def secondLevel = mavenRepo.module("org.gradle", "secondLevel", "1.0").dependsOn(thirdLevel).publish()
-        def firstLevel = mavenRepo.module("org.gradle", "firstLevel", "1.0").dependsOn(secondLevel.groupId,
-            secondLevel.artifactId, secondLevel.version, null, null, null, [[groupId: '*', artifactId: '*']]).publish()
+        repository {
+            'org.gradle:thirdLevel:1.0'()
+            'org.gradle:secondLevel:1.0'() {
+                dependsOn 'org.gradle:thirdLevel:1.0'
+            }
+            'org.gradle:firstLevel:1.0' {
+                dependsOn(group:'org.gradle', artifact:'secondLevel', version:'1.0', exclusions: [[groupId: '*', artifactId: '*']])
+            }
+        }
 
         and:
         buildFile << """
-repositories { maven { url "${mavenRepo.uri}" } }
-configurations { compile }
 dependencies {
-    compile "$firstLevel.groupId:$firstLevel.artifactId:$firstLevel.version"
+    conf "org.gradle:firstLevel:1.0"
 }
 """
+
+        repositoryInteractions {
+            'org.gradle:firstLevel:1.0' {
+                expectGetMetadata()
+                expectGetArtifact()
+            }
+            'org.gradle:secondLevel:1.0' {
+                expectGetMetadata()
+                expectGetArtifact()
+            }
+        }
 
         expect:
         succeeds "checkDep"
